feat(pdf command): add option to change pdf viewer + fix readme
This adds the config option `vim.g.openscad_pdf_command`, which allows
for the use of pdf viewers other than zathura, it also removes the
healthcheck for zathura and sets the default viewer to ''.

The default viewer being '' might cause problems for people who
currently use zathura, as it would break their setup. However, trying to
view the manual without setting a pdf viewer causes a warning, so I
think it's fine.

This pr also fixes an issue where the readme claims that
`vim.g.openscad_help_manual_trig_key` is the config value for the manual
opening keybind, when it is actually `vim.g.openscad_manual_trig_key`.